Add Jenkins Pipelines support to Option Provider
The Jenkins Pipeline plugin introduces the need to use more generic
classes in certain cases when working with Pipelines and Multibranch
Pipelines vs other types of Jenkins projects.

For the Jenkins Rundeck Option Provider, a few uses of AbstractProject
needed to be changed to the more generic Job class, as described in
the Basic update section of the Jenkins Pipelines Plugin Developer
Guide (see the Source reference below).

To aid with determination of Option Provider URLs, the error message
in the case of an invalid project parameter in the Option Provider URL
has also been enhanced to display a complete Jenkins item tree showing
the full name of each item, which can then be used as a project
parameter. This is especially helpful for Jenkins Multibranch
Pipelines which contain a partially URL encoded path.

Source
https://github.com/jenkinsci/pipeline-plugin/blob/master/DEVGUIDE.md#user-content-basic-update